[ME] Simplify transaction adapters and consolidate Studio loading
