+++ src/MaterialEditor.Core.Studio/Core.MaterialEditor.Studio.projitems
@@ -11,7 +11,6 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.SceneController.Persistence.Copy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.SceneController.Persistence.Properties.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.SceneController.Persistence.Textures.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.SceneController.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.SceneController.Edits.Core.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.SceneController.Edits.Cubemap.cs" />
