+++ src/MaterialEditor.Core/Core.MaterialEditor.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>Core.MaterialEditor</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Persistence\MaterialCubemapImportTransaction.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.CharaController.Restore.Properties.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Persistence\MaterialEditLoadContext.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.CharaController.Persistence.Properties.cs" />
